Add Card.ActiveProtocol accessor
This piece of information is useful for some things (such as whether GET
RESPONSE APDUs are needed and such).

The active protocol can already be obtained with Card.Status(), however
we already have it stashed in the struct (returned by SCardConnect()),
so add a getter to make it slightly faster and error-free.